Team done, not tested

+++ src/controllers/teams.ts
@@ -21,9 +21,25 @@ try {
             reject(error)
         return;
         }
+teamName = teamName.toLowerCase();
 
+TeamsModel.findOne({ name: teamName,
+  user_id: userId})
+  .then((exist: any)=>{
+if(exist){
 
-TeamsModel.create({
+  let error: RESPONSE_TYPE = {
+    data:[],
+    message: "Team already exist.",
+    status: 409,
+    statusCode: "RESOURCE_ALREADY_EXIST"
+}
+reject(error)
+return;
+}
+else{
+
+  TeamsModel.create({
     name: teamName,
     description: description,
     user_id: userId
@@ -66,9 +82,42 @@ TeamsModel.create({
     return;
 })
 
+
+}
+
+
+  })
+  .catch((err: any)=>{
+
+
+    let error: RESPONSE_TYPE = {
+      data:[],
+      message: "Team creation failed",
+      status: 400,
+      statusCode: "UNKNOWN_ERROR"
+  }
+  reject(error)
+  let err_found: ErrorDataType = {
+      msg:` ${err.message}` ,
+      status: "STRONG",
+      time:   new Date().toUTCString(),
+      stack:err.stack,
+      class: <string> <unknown>this
+
+
+  }
+
+
+  LogError(err_found);
+  return;
+  })
+
+
+
       }
           
 catch(err: any){
+
     let error: RESPONSE_TYPE = {
         data: [],
         message: "Failed to update job data.",
@@ -98,11 +147,88 @@ catch(err: any){
       });
     }
     
+ // Check if a team exists
+ getATeam(teamId: ObjectId): Promise<RESPONSE_TYPE> {
+  return new Promise((resolve, reject) => {
+    TeamsModel.findById({_id:teamId})
+      .then((team) => {
+        if (team) {
+          const response: RESPONSE_TYPE = {
+            data: [team],
+            message: "Team exists",
+            status: 200,
+            statusCode: "SUCCESS"
+          };
+          resolve(response);
+          return;
 
+        } else {
+          const response: RESPONSE_TYPE = {
+            data: [],
+            message: "Team does not exist",
+            status: 404,
+            statusCode: "RESOURCE_NOT_FOUND"
+          };
+          resolve(response);
+          return;
+
+        }
+      })
+      .catch((err) => {
+        const error: RESPONSE_TYPE = {
+          data: [],
+          message: "Failed to check team existence",
+          status: 500,
+          statusCode: "UNKNOWN_ERROR"
+        };
+        reject(error);
+        return;
+      });
+  });
+}
+
+getATeamByName(name: string): Promise<RESPONSE_TYPE> {
+  return new Promise((resolve, reject) => {
+    TeamsModel.find({name})
+      .then((team) => {
+        if (team.length>0) {
+          const response: RESPONSE_TYPE = {
+            data: team,
+            message: "Team exists",
+            status: 200,
+            statusCode: "SUCCESS"
+          };
+          resolve(response);
+          return;
+
+        } else {
+          const response: RESPONSE_TYPE = {
+            data: [],
+            message: "Team does not exist",
+            status: 404,
+            statusCode: "RESOURCE_NOT_FOUND"
+          };
+          resolve(response);
+          return;
+
+        }
+      })
+      .catch((err) => {
+        const error: RESPONSE_TYPE = {
+          data: [],
+          message: "Failed to check team existence",
+          status: 500,
+          statusCode: "UNKNOWN_ERROR"
+        };
+        reject(error);
+        return;
+      });
+  });
+}
 
 
      // Update team information
-     updateTeam(teamId: ObjectId, newTeamInfo: TeamUpdate) {
+     updateTeam(teamId: ObjectId, newTeamInfo: TeamUpdate): Promise<RESPONSE_TYPE>  {
         return new Promise((resolve, reject) => {
           try {
             // Your code to update team information goes here
@@ -128,9 +254,9 @@ let error: RESPONSE_TYPE = {
             } 
 
 
-            TeamsModel.updateOne({_id: teamId}, {$set: newTeamInfo})
-            .then((data)=>{
-            let response = {
+            TeamsModel.updateOne({_id: teamId}, newTeamInfo)
+            .then((data: any)=>{
+            let response: RESPONSE_TYPE = {
               data: [],
               message: "Team updated successfully",
               status: 200,
@@ -168,7 +294,7 @@ let error: RESPONSE_TYPE = {
  
           } catch (err: any) {
             // Handle any unexpected error during team update
-            let error = {
+            let error :  RESPONSE_TYPE = {
               data: [],
               message: "Failed to update team",
               status: 500,
@@ -192,7 +318,7 @@ let error: RESPONSE_TYPE = {
       }
     
       // Fetch teams created by a user
-      fetchTeamsByUser(userId: ObjectId) {
+      fetchTeamsByUser(userId: ObjectId): Promise<RESPONSE_TYPE>  {
         return new Promise((resolve, reject) => {
           try {
             // Your code to fetch teams by user goes here
@@ -200,8 +326,8 @@ let error: RESPONSE_TYPE = {
             // Resolve the promise with a success response
 
             TeamsModel.find({user_id: userId} )
-            .then((data)=>{
-                let response = {
+            .then((data: any)=>{
+                let response :  RESPONSE_TYPE= {
                     data,
                     message: "Teams fetched successfully",
                     status: 200,
@@ -213,7 +339,7 @@ let error: RESPONSE_TYPE = {
             .catch((err: any)=>{
 
 
-                let error = {
+                let error :  RESPONSE_TYPE= {
                     data: [],
                     message: "Failed to fetch teams",
                     status: 500,
@@ -242,7 +368,7 @@ let error: RESPONSE_TYPE = {
 
           } catch (err: any) {
             // Handle any unexpected error during team fetch
-            let error = {
+            let error:  RESPONSE_TYPE = {
               data: [],
               message: "Failed to fetch teams",
               status: 500,
@@ -263,49 +389,16 @@ let error: RESPONSE_TYPE = {
       }
     
       // Delete a team
-      deleteTeam(teamId: ObjectId) {
+      deleteTeam(teamId: ObjectId, userId: ObjectId) : Promise<RESPONSE_TYPE> {
         return new Promise((resolve, reject) => {
-         /* 
-            try {
-            // Your code to delete a team goes here
-            // You can use the provided teamId
-            // Resolve the promise with a success response
-            let response = {
-              data: [],
-              message: "Team deleted successfully",
-              status: 200,
-              statusCode: "SUCCESS"
-            };
-            resolve(response);
-          } 
-          
-          catch (err) {
-            // Handle any unexpected error during team deletion
-            let error = {
-              data: [],
-              message: "Failed to delete team",
-              status: 500,
-              statusCode: "UNKNOWN_ERROR"
-            };
-            reject(error);
-            let err_found = {
-              msg: `${err.message}`,
-              status: "STRONG",
-              time: new Date().toUTCString(),
-              stack: err.stack,
-              class: <string><unknown>this
-            };
-            LogError(err_found);
-          }
-
- */
+       
 
 try{
           
-          TeamsModel.deleteOne({_id: teamId}  )
+          TeamsModel.deleteOne({_id: teamId, user_id: userId}  )
           .then((data)=>{
-              let response = {
-                  data,
+              let response :  RESPONSE_TYPE= {
+                  data: [],
                   message: "Teams deleted successfully",
                   status: 200,
                   statusCode: "SUCCESS"
@@ -316,7 +409,7 @@ try{
           .catch((err: any)=>{
 
 
-              let error = {
+              let error:  RESPONSE_TYPE = {
                   data: [],
                   message: "Failed to delete teams",
                   status: 500,
@@ -339,7 +432,7 @@ try{
 }
 catch (err: any) {
     // Handle any unexpected error during team fetch
-    let error = {
+    let error:  RESPONSE_TYPE = {
       data: [],
       message: "Failed to fetch teams",
       status: 500,
@@ -366,4 +459,5 @@ catch (err: any) {
 
   }
   
-  module.exports = TeamModule;
+  
+  export default new TeamModule();
